organic ai/requirements.txt
# Core dependencies
python-telegram-bot[all]==20.7
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0

# Database
aiosqlite==0.19.0

# Utilities
pytz==2023.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Type stubs
types-pytz==2023.3.1.1
types-aiofiles==23.2.0.0
types-requests==2.31.0.20240311 